BUG: fixing indexing

+++ gsxform/scattering.py
@@ -233,8 +233,8 @@ class Spline(ScatteringTransform):
         E, V = compute_spectra(self.W_adj)
         eig_max = torch.max(torch.diag(E))
 
-        x1 = torch.diag(E)[np.floor(self.n_nodes / 4)]
-        x2 = torch.diag(E)[np.ciel(3 * self.n_nodes / 4)]
+        x1 = torch.diag(E)[np.floor(self.n_nodes / 4).astype(np.int)]
+        x2 = torch.diag(E)[np.ciel(3 * self.n_nodes / 4).astype(np.int)]
 
         # compute wavelet operator
         psi = spline_wavelets(
